+++ benchmarks/flutter_cache_benchmark/android/app/src/main/kotlin/com/example/flutter_cache_benchmark/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.cache
+package com.example.flutter_cache_benchmark
 
 import io.flutter.embedding.android.FlutterActivity