+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.3
+
+* **io**: `HttpFilesCache` remove isolate; add concurrent IO; update benchmarks
+* **web**: replace `HttpHiveCache` by `HttpWebCache` which is created from scratch
+
 ## 2.0.2
 
 * Fix web compatibility analysis issues